tdd: Fix post-test-suite terminal input glitching
The spec setup for examples concerning the `--safe` option
was creating a problem where after running the test suite,
the first couple keystrokes in the terminal window were ignored
(or, presumably, consumed by the background thread
in which `tail -f` was running).

This commit fixes this issue by adding a `&` to the `tail -f` command.
I'm not 100% certain why this works, but presumably,
`&` divorces the process from STDIN?