bfx: internalize terminal-size gem dependency
The `terminal-size` gem presents an error[0] on some systems:

    >> require 'terminal-size'
    <internal:.../kernel_require.rb>:160:in `require': cannot load such file -- terminal-size (LoadError)

This bug is caused by faulty file permissions in the `terminal-size`
gem, as published on rubygems.org:

    $ gem install terminal-size
    $ ls -l $(gem which terminal-size)
    -rw-------   1 root    root   1451 May  1 18:19 terminal-size.rb

If the user running ruby is different from the one who owns the file,
then `require 'terminal-size'` will fail.

This commit addresses the problem by simply copying the entire,
one-file library in a `/vendor` directory and requiring that,
as opposed to using the RubyGems package management system.

[0]: https://github.com/rking/terminal-size/issues/4