dev: Prefer global vars to constants for stdio streams
As explained in this SO answer[0],
$stdout and $stderr can be redefined at run-time,
whereas STDOUT and STDERR cannot be changed
(the latter two being the default values of the former two).

This commit has no practical impact for the time being,
but may facilitate debugging in the future
by allowing stdout / stderr to be redefined within Ruby,
rather than simply redirected in the shell.

(Might also be a case of YAGNI,
but also provides consistency with 3c01b2e.)

[0]: https://stackoverflow.com/a/6671770/4865822