bfx: Prefer metadata- over filename-timestamps
Previously, the application gave equal weight
to metadata and filename timestamps,
simply preferring the one that came earlier.

This created issues for filenames with a date but no time:
`Time.parse` returns a timestamp for midnight of that day,
which is virtually always earlier than the actual timestamp.

There being no good way to automatically distinguish
between good and bad file timestamps,
it was decided that metadata timestamps should never be overridden.