bfx: Fix file extension when optimizing media
Using the `--optimize-for={desktop,web}` options adds an extra step
where the optimized photo/video is written to a tempfile
and then moved to the destination.

The `#dest_path` method uses the same extension as the source file,
but optimization can sometimes involve changing the file extension
(e.g., .mov -> .mp4 / .heic -> .jpg),
causing an optimized .jpg tempfile to be renamed as an .heic.

This commit fixes this issue.