ftr: Remove --volume CLI option
archivist was originally built with a --volume CLI option
because it was envisioned as a utility script
to be launched via systemd:

    # dcim-transfer.service
    [Unit]
    Description=Digital Photo Transfer for Ricoh GR
    BindsTo=dev-disk-by\x2dlabel-RICOH_GR.device
    After=dev-disk-by\x2dlabel-RICOH_GR.device

    [Service]
    ExecStart=archivist --volume=/media/ricoh-gr --source=<...> --dest=<...>
    Restart=no

    [Install]
    WantedBy=dev-disk-by\x2dlabel-RICOH_GR.device

The idea was for the `ExecStart` command
to automatically handle mounting and unmounting of the target volume--
that is, to wrap the entire transaction into a single command.

As archivist grows to support more generalized use cases,
the --volume option has been deemed extraneous to its core purpose
of organizing/renaming files according to their timestamp metadata
and optimizing them for various library types.

To achieve the original goal without the --volume option,
simply wrap your `ExecStart` line in a call to `sh -c`:

    ExecStart=sh -c 'mount /media/ricoh-gr; archivist --source=<...> --dest=<...>; umount /media/ricoh-gr'