rfg: Add dedicated method for dest. file extension
When archivist is asked to optimize media prior to import
(i.e., transcode videos at a lower bitrate
or downscale photos to a lower resolution),
the operation is not atomic, so it writes to a tempfile
and then moves the tempfile into place.

In many cases, the optimization process
produces a filetype different from the source file.
For instance, HEICs may be converted to JPEGs
for portability / compatibility with platforms
that do not contain HEIF support yet.

Prior to this commit,
the logic for determining the appropriate destination file extension
was unclear and scattered.
Now, it is encapsulated in the #dest_extname method.

Bonus: This commit also removes the Archivist::Extname class,
which was arguably a premature optimization and a case of YAGNI.