bfx: (archivistd) Keep empty directories in inbox
Originally, archivistd would slurp files
from the inbox into the staging directory
and the remove all the empty subdirectories left over.

This keeps the directory structure clean,
but archivistd will fail to detect new files in those subdirectories
if they are respawned/regenerated later on.
This is because under the hood,
archivistd uses inotify to detect new files,
and inotify is not recursive.
Thus, archivistd explicitly specifies a separate "watch"
for every single subdirectory of the inbox at launch.

This commit addresses this issue by removing the logic
for cleaning up empty subdirectories of the inbox.