bfx: Remove check for presence of optipng binary
This commit removes a call to `Optipng.available?`,
which raises an error on Alpine Linux (i.e., in Docker containers)
because it depends on the whereis UNIX utility:

    /usr/local/bundle/gems/pipe-run-0.3.0/lib/pipe-run.rb:34:in `popen': No such file or directory - whereis (Errno::ENOENT)
        from /usr/local/bundle/gems/pipe-run-0.3.0/lib/pipe-run.rb:34:in `run'
        from /usr/local/bundle/gems/command-builder-0.3.2/lib/command-builder.rb:229:in `execute'
        from /usr/local/bundle/gems/unix-whereis-0.1.1/lib/unix/whereis.rb:87:in `file?'
        from /usr/local/bundle/gems/unix-whereis-0.1.1/lib/unix/whereis.rb:168:in `available?'

whereis is not provided by default in Alpine Linux,
and requires the `util-linux` package.

A safety check for the presence of an optipng binary would be nice,
but we don't check for the presence of ffmpeg or imagemagick either,
so in the interest of consistency, we're just getting rid of this one.