ftr: Enable configuration natively within Ruby
Previously, photein was intended
to only be run from the shell (i.e., as an executable),
and so configuration was only accepted via OptionParser.

Since photein's sister utility xferase
is being rewritten to use photein internally as a Ruby library
rather than invoking it in a system subprocess,
it needs to be able to set this configuration directly in Ruby code.
For that reason, this commit adds a `Photein::Configuration.set` method
to set it directly.

It also comes with a version increment.