ftr: Support multiple destination libraries
This commit is a significant rewriting of Photein's core import logic.
Where previously, a user would specify a single destination directory
and a separate, single optimization level (desktop or web),
photein now accepts three separate destination arguments,
each corresponding to a separate optimization level:

    # Before
    $ photein --src [SOURCE] --dest [DEST] \
        --optimization-level [desktop|web]

    # After
    $ photein --src [SOURCE] \
        --library-master [DEST-1] \
        --library-desktop [DEST-2] \
        --library-web [DEST-3]

The purpose of this change was to support
the simultaneous (multi-threaded) importing of individual files
to multiple destinations.
This feature will be leveraged in a subsequent release of
Photein's sister utility, Xferase.