bfx: Simulate case-insensitive file globbing
According to the offical Ruby documentation,
the special flag I used to use to make `Dir.glob` case-insensitive
was never actually supposed to work:

> Note that the pattern is not a regexp, it’s closer to a shell glob.
> See File::fnmatch for the meaning of the flags parameter.
> Case sensitivity depends on your system (File::FNM_CASEFOLD is ignored).

In my very hazy recollection, it used to work on Linux before--
but it certainly doesn't work anymore.

This commit switches to a regexp instead to support case-insensitivity.